requirements.txt
streamlit==1.25.0
tensorflow==2.12.0
numpy==1.24.0
pillow==9.5.0
